Rename AF_FINISH_FILES()'s internal variables
This patch is the second in a series to clarify the terminology and prepare
separation of the build-time finishing framework into its own project.

We consider the build-time finishing framework being an extension of the
GNU Autotools. Consequently, we rename variables
  - libsh_cleanfiles         to af__CLEAN_FILES and
  - libsh_config_status_deps to af__config_status_deps,
mimicing automake(1)'s naming conventions. Both of them get ultimately
AC_SUBST()ed into Makefile.

Following the same conventions, we also rename variables
  - libsh_gensubst   to af__gensubst and
  - libsh_finalizees to af__unfinished.
The latter is also done for clarifying terminology.