Rename output variable FINALIZE to FINISH
This patch is the third in a series to clarify the terminology and prepare
separation of the build-time finishing framework into its own project.

It renames output variable FINALIZE to FINISH (because that's what it does
now). For claryfing terminology in auto/gensubst, variable finalizees is
also renamed to unfinished.