gensubst: Add 'quote' substitution
The new substitution applies multiple levels of RE/RS quoting (with va-
rying DELIMITERs) to a STRING.

EXAMPLE:

STRING is used for matching in some sed(1) 's|||' command, yet that very
command is constructed in some sed(1) 's///' command. This is achieved by

    rs_re_string=`gensubst quote "$string" re \| rs /`

or just (the last DELIMITER is '/' by default)

    rs_re_string=`gensubst quote "$string" re \| rs`

instead of

    re_string=`gensubst quote-re "$string" \|`
    rs_re_string=`gensubst quote-rs "$re_string"`

Thus,

    string=./config.status

becomes

    re_string='\./config\.status'

('.' always special in RE, but '/' not as '|' is the DELIMITER),
which becomes

    rs_re_string='\\.\/config\\.status'

('\' always special in RS, and '/' special as it's the DELIMITER),
yielding the desired effect (i.e., '\' and '/' unquoted by 's///',
'.' staying quoted for matching in 's|||').