gensubst: Add 'expand' substitution
The new substitution reads all variable assignments from the (top-level)
makefile and (recursively) expands '$(NAME)' or '${NAME}' substitutions
contained in the specified STRINGs. The expansion is controlled by spe-
cifying a 'mode=MODE' pair.

Example:

    SED=sed gensubst expand mode=eval -- 'SED=$(SED)'

yields the configured value of output variable @SED@, supplying the de-
fault 'sed' from the environment if necessary, i.e. outputs

    SED='/bin/sed'

or

    SED='sed'

which can be used within 'eval "`...`"' to retrieve the result of

    AC_PROG_SED

This is a more than appropriate replacement for the awkward 'config.sh'
mechanism.