dl_warn(): Work around nasty (GNU) sed(1) regression
Up to some sed(1) version before 4.7, the following was possible:

  /^\(foo\)bar/{;s//\1lish/; ... }

In particular: it was perfectly OK to backreference to a previous match
within the right-hand side of an 's///' statement as long as its left-
hand side was empty (i.e., "reused" the previous match).

Alas, this is no more! Therefore, work around this regression by
  - removing the parentheses from the match (they're useless now) and
  - rematching by the 's///' statement's left hand-side in a parenthesized,
    yet simplified form so we may backreference to that within the
    right-hand side.

Hmm, while pondering the issue once more, I came to the conclusion that
the true culprit is the following 's/$/\1meta.../' statement! 8-O
So sed(1) is right now and *I* had been wrong?!
Anyway, work around it by also merging the two 's///' statements into one.

Moreover as an optimization, the s/// statement within the initial loop
is moved behind the loop and made to operate globally.